[iOS] Enable the clearing of browsing data in web
This CL enables the feature by default.
It also runs directly the closure instead of posting it to have the
sequence_checker validated.

It also removes the observation of the browsing data removal in the
WebController. This is because disabling the WebUsage is supposed to
be tied to the spinner displayed on browser.
We should find a way of synchronizing both of them.

Bug: 981375
Change-Id: Iac47974e0d34f6977b2879dc00cd049e3950e387
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1847308
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#708580}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33a49c9dfaf76ca54b16197b5bb1c46647f303a4