[iOS][EG2] Add a common method for rotation
This adds a common method to our testing class that can be used in EG1
and EG2.
|EarlGrey| has 2 methods to rotate the device, once for EG1 and another
for EG2. Which makes this check pretty common in the EG2 migration.
Adding it to the base test class instead of to |EarlGrey| to avoid
creating a 3rd method signature for it. Once EG1 support is removed a
simple search and replace can be used to clean this.

Bug: 1016368

Change-Id: I24b2c4ae56266111e6069990252164090df80bb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1871874
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#708696}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 254044d90910c2116a4231327a2df99dccc9d473