[iOS][EG2] Create keyboard app interface
Create an app interface for the app code in fallback_coordinator_egtest
related to the keyboard.
Move the code from the test to this interface so it can be used later
to migrate the test to EG2.

Bug: 1017175
Change-Id: I08a85c003a1797d2ef5fa0cd8e4ec444dd05fb3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1876392
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#708952}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dcf155ffd09ad1b61a4670e34d8a1bc491d1ee0e