[ios] Fix BadgeViewController interpretation of badgeState
Since BadgeState is a mask, a direct equality comparison to
BadgeState values for the BadgeItem properties is not the correct
way to assess if a badge is accepted.

To add testing support for this case, accepted a11y identifiers
for BadgeButtons are created, and they will be properly set
in setAccepted: in BadgeButton.

Bug: 976901, 1016793
Change-Id: Id17eb82e0764fbbf80b871df0b974fe148c2339f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1872938
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#709037}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4fcce6d720daa4ffd8e8eceb5332d1f05383f568