[iOS] Log when existing URL is being opened
The idea is to log when the user is opening a URL (from another app)
which is already opened in the app.
Having this metrics will allow us to know if it is impactful to handle
this special case.

Bug: 1014964
Change-Id: I94c7b7810eb66c81ec310057d1b4b991dad55ab7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1869412
Reviewed-by: Nik Bhagat <nikunjb@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#710130}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 470af3323241bab615b5151adebfe238c19b5070