[iOS] Plumb Browser through remainder of auth flow.
This is the final of five CLs breaking crrev.com/c/1860025 into smaller
chunks.

This CL updates the classes in ui/authentication to accept a Browser
instead of a BrowserState, and updates their callers, who at this point
all are created with Browsers to pass in.

The workarounds in the parent CL are removed, and now all of the inits
for SettingsNavigationController use Browser.

Bug: 1018746
Change-Id: Iaa6b20279daa64d377730c74c29b14f25c9d1404
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1883724
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#710332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 67bcad56515a70e7aca0353574f8f621c56d335c