[iOS] Fix bug where omnibox popup row selection color was incorrect
This bug only occurs on iOS 12. The root cause is that, on iOS 13 we
use the dark mode apis to color things in incognito. This means all
colors can be set once and incognito can be changed afterwards and the
colors will update. For iOS 12, the colors are static, so they must be
set after the incognito status is known.

Bug: 1017138
Change-Id: I8084646b07d2bcbf8b48d91b0d7d341a8fb9bece
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1886684
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#710399}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9445030a83995ee9995cc98e6391b4c905180ce3