[ios] Removes ReportAnIssue presentedViewController DCHECK
The DCHECK is triggered whenever Report an Issue is tapped and an
Infobar Message banner is currently being presented.

This DCHECK makes an assumption about another VC, and its misleading
since baseViewController might not even be the VC that ends up presenting
report an issue. Which means that this doesn't give us certainty that
the presenterVC is not presenting another VC.


This works as expected since BVC will dismiss the InfobarBanner and
then present report an issue.

Bug: 1016962
Change-Id: I516a6513422fe8817b291e00f0944d602b98357b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1887754
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#710800}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 702af1e55231fe606a26a5d11202bc9bf5342dbf