[iOS] Make net:test_support dependencies in egtests explicit.
ChromeTestCase exposes an EmbeddedTestServer property for use by tests,
and includes the header for that class in its own header. This means
that tests can (and do) use symbols from EmbeddedTestServer without
explicitly including them. To make this work, //net:test_support is
a public dependency of the egtest test_support targets, so the individual
egtest targets can link correctly.

This is in general not in line with our "include what you use" policy,
which would prefer that code which uses symbols from an API directly
includes the header for that API, and explicitly declares it as a dep
in gn.

This CL:

  - Makes the test_support dependencies on //net:test_support private.
  - Changes the EmbeddedTestServer header include in chrome_test_case.h
    to be a forward declaration.
  - Updates nineteen egtest.mm files to explicitly include the
    EmbeddedTestServer header.
  - Updates twenty-seven targets in BUILD.gn files to have add a
    //net:test_support as a dependency.


Change-Id: Iea336c66a1a7423732e3aaedcf865882b7ac3a0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1892814
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#712100}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb0a29d100e75a404c4949cd3ce6aebf4f977d5e