[iOS] Turn off tab grid animation when Voiceover is on
When the animation is on, there is a bug where if the user swipes
left-to-right to go to the next Voiceover view during the animation,
the Voiceover cursor ends up on view from the previous tab grid page,
instead of the next tab grid page. This happens because the previous
tab grid page's views are still accessible during the animation.

To prevent this, the animation can be turned off when Voiceover is
enabled. This means that when the user selects a new tab, the views
immediately update.

Bug: 978673
Change-Id: Id4b64f391614b8719e3f5d5d2e924df7d8983123
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1906348
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#715239}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 297c73f88d62ff6c1057a882504085c89c247740