[iOS] Move content below header after rotation when scrolled to top.
When the expanded toolbar height is updated while the page is scrolled
to the top, the top of the page will be obscured behind the toolbar.
Toolbar heights only change during trait collection updates, so the
page can be re-scrolled to the top after the trait collection is
finished updating if it was detected that it was scrolled to the top
when the update began.

Bug: 1017146
Change-Id: I1065cae2c1824e61a8688b845b88211013fa1949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1914968
Auto-Submit: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#715347}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7dbefa699e9c67ddb02136532d99cd278124df32