[Autofill Auth UI] Rename webauthn dialog flow
Reason for this CL: We did implementation review with Chrome UX and
one feedback was the verify pending dialog was too short for a modal
dialog. Therefore UX decision is for this dialog, we also reuse the
template of the webauthn dialog (adding an illustration on top). So
it makes sense to merge the separate MVCs of webauthn_offer_dialog
and verify_pending_dialog into one, and show different content for
different states (in other words, adding kVerifyPending as a new state
in WebauthnDialogModel).

This is a renaming CL which is the precursor of the major merging CL.

Includes two changes:
1. Rename all WebauthnOffer- to Webauthn- to avoid confusion;
2. In AutofillClient, move all webauthn related functions under build
flags to ensure these functions are all on desktop.

No other change on flow logic.

TBR=michaelbai@chromium.org,mahmadi@chromium.org,martinkr@google.com

Change-Id: I80ae00d219cdaf868056a04938c6968ed186ad3d
Bug: 949269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1925761
Commit-Queue: Siyu An <siyua@chromium.org>
Reviewed-by: Evan Stade <estade@chromium.org>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#718394}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0ad78f7655d513de3a834b13a0c9b0d224e36434