[iOS] Fix dark mode visual bug on signin screen
The root cause is that, because CALayer uses CGColors, which don't
support the whole dynamic color infrastructure, we have to manually
set these colors to the correct version in -traitCollectionDidChange.
However, CALayer automatically animates any change to these properties.

This means that when the color mode changes, the rest of the colors
update immediately, but the layer's colors animate to the correct color
over the next ~0.5 seconds.

To fix this, we can use CATransaction and setDisableActions to disable
the automatic animation.

Fixed: 1028091
Change-Id: Id217c5387bffc77a5a971aac0e43198a1ffa8200
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1940328
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#719585}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5b12c493d15c9e801e0d4bd680d7c8075802fb6d