[iOS] Use more inclusive language for allow/block lists.
This CL removes most of the uses of 'blacklist' and 'whitelist' in //ios/.

"whitelist" is changed to "allow list", or in some cases just "list".
"whitelisted" is changed to "allowed" or "allow-listed" depending on context.
"blacklist" is changed to "block list" or "blocked list".
"blacklisted" is changed to "blocked".

All of the changes to comments and internal-to-iOS identifiers should be safe.

Two strings have their names changed.

Two categories of usage of these terms were not changed:

- Method calls from components code. NTP tiles, passwords, and Translate are
  the sources for most (all?) of these.

- Public methods in ChromeWebView, which consumers may be using.

Bug: 842296
Change-Id: I7483a954457325f554a7abf7179faafba30c5e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1936254
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#719689}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e1cbb48ec63cdf36801774207a958f10f28e754e