[iOS] Freeze Safari version in Mobile UserAgent
This CL freezes the Safari version in the Mobile User Agent, aligning with
https://blog.chromium.org/2019/05/improving-privacy-and-security-on-web.html.

It will also increase compatibility with WebKit, as it is frozen in
WebKit mobile User Agent.

The iOS 12 version was introduced in http://crrev.com/c/1102485.
This CL was made during the beta version of iOS 12, my guess is that the
version was frozen in WebKit before the Stable version was released, so we
released a wrong version by mistake.

Bug: 1026615
Change-Id: I7641bc54327f3d135501dc47c365c4458ca8dace
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1926488
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Aaron Tagliaboschi <aarontag@chromium.org>
Reviewed-by: Peter Lee <pkl@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#720198}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bd1c791ef29b05666efafc8957745ac9f2b8ec19