Collect user opt-in before enabling account-storage for passwords
This CL adds a pref that stores which accounts have opted in to use the
Google account storage for passwords. While this pref is false,
passwords can not be saved to the account, and passwords in the account
are not downloaded.
It also adds an entry to the password autofill dropdown that lets the
user set the opt-in.

Bug: 1024332
Change-Id: Iac552c310d0e9f19ee2bb1893e217e60b1114d31
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1645294
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#720206}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c3c374663b69924fee30dec66c36bf1a2397d3cd