Release all ChromeIdentity objects when forgetting them.
This CL ensures that all ChromeIdentities objects are not retained by
the global autorelease pool when they were forgotten.

This is needed as relying on the global autorelease pool has as a side
effect the fact that the underlying GTMSessionFetcher objects was also
retained and this confused the EarlGrey infrastructure that believed
that a network request was still on-going during the tear down phase.

Bug: 1011798
Change-Id: Ica44d3a26d4f37d30928e9a53c104aed7bd59e13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1948436
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#721031}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 190fdbf1ac8b06743c1de12fb96d382004c31801