Make progress_indicator_egtest.mm compatible with EG2
This adds the file as an EG2 target, updates #includes, and replaces
calls to helpers that can't be used in EG2.

This also adds a ProgressIndicatorAppInterface class that provides
a matcher for MDCProgressViews with a given progress value.

Bug: 987646
Change-Id: I88b8bc9496bf962acbba59d855205cf19cf97fd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1949266
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#721484}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a284b88def777d5380357f754444f84ca816f71a