[iOS] Request desktop by default
This CL adds the logic for requesting the desktop version of web sites
by default when kDefaultToDesktopOnIPad is enabled.
It also adds the logic for when to request desktop/mobile based on
the size classes.
The user agent types are used for three things:
1. Updating the user agent sent to the web page.
2. Storing the user agent used for back/forward navigation and
   knowing what request (mobile or desktop) to display in the menus.
3. Inheriting from one item to the other.

1 and 2 should be kept in sync. 3 can be different.
So this CL introduces a UserAgentType used for inheritance, which can
be different from the one used for updating the user agent.

Bug: 1025225
Change-Id: Iad219985ec2625979e2cc7f2aa0f1eb4dbfacec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1919352
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#721965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a63db37851b53af164babdb8ce67a0f0800b188