[ios] Update languages before executing Translate from the modal
This change ensures that any change in the source/target language by the
user in the Infobar modal will be saved and changed in the InfobarTranslateDelegate
right before executing a new Translate. It also cleans up some logic to ensure
that the correct new language is always displayed in the modal and also the correct
new language is checked in the LanguageSelectionViewController.

Bug: 1014959
Change-Id: I4bd40c1b916e0479d551bd5962231916a04a920d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1952424
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#722358}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07a96a59449af2ed4366b7d37a8583b581e1600e