Fix CanGoBack check when creating loud SB errors
Checking CanGoBack when constructing an interstitial from a throttle
(the way they are created with committed interstitials) resulted in
false negatives in some cases. This functionality is not used in Chrome
but is in Android Webview for safe browsing interstitials (they do not
show the Back To Safety button if there is nothing to go back to).
This CL introduces a separate method to check in that case.

Test: Covered by testSafeBrowsingDontProceedNavigatesBackForMainFrame

Bug: 1031620
Change-Id: I3638e173dcc226a18110153c6a33595a5cf45ab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1956120
Commit-Queue: Carlos IL <carlosil@chromium.org>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Auto-Submit: Carlos IL <carlosil@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#722703}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3d401f53fa7b735e06961d07ebb4f3f8ec5c9744