[Passwords] Remove PasswordForm.preferred field
Before this patch:
PasswordManaged maintained a Boolean "preferred" field that would
indicate the preferred username by a user for a specific website.

- When filling, the one that with the preferred username is filled
(together with the corresponding password).
- After successful login, the used  username is marked "preferred",
and all other usernames are marked not preferred.

In addition, the "date_last_used" is populated upon successful login
with the current time.

This CL deprecates the "preferred" field and switches to using
the "date_last_used" instead.
such that when filling, the most recently used username is filled.

Not only this reduces the number of records to be saved upon login,
but this will simplifies the logic for filling and saving.
Such simplification is important for the upcoming butter for passwords
efforts.

Change-Id: I6d7575dde289d38632859d617e07df9703992792
Bug: 999949
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1741809
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Auto-Submit: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#722994}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 04c4f36233cb2ab4cda2aca96fb79f5f85b790f7